backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
web3==7.6.0
python-dotenv==1.0.1
google-genai>=1.7.0
aiosqlite==0.20.0
pydantic==2.9.0
pydantic-settings==2.6.0
websockets==13.0
httpx>=0.28.1
apscheduler==3.10.4
vaderSentiment==3.3.2
